+++ projects/pastanaga/src/lib/textfield/password-input/password-rule.ts
@@ -1,0 +1,23 @@
+export interface IPasswordRule {
+    id: string;
+    label: string;
+    regexp?: RegExp;
+    length?: number;
+    isValid?: boolean;
+}
+
+export class PasswordRule {
+    id: string;
+    label: string;
+    regexp?: RegExp;
+    length?: number;
+    isValid: boolean;
+
+    constructor(data: IPasswordRule) {
+        this.id = data.id;
+        this.label = data.label;
+        this.regexp = data.regexp;
+        this.length = data.length;
+        this.isValid = !!data.isValid;
+    }
+}